+++ poke-gambler/src/App.js
@@ -1,17 +1,8 @@
 import logo from './logo.svg';
 import './App.css';
-import { Container, Card, Icon, Button } from 'semantic-ui-react';
+import { Container, Button } from 'semantic-ui-react';
+import Wallet from './components/Wallet'
 
-const extra = (
-  <a>
-    <Icon name='bitcoin' />
-    0.000018
-    <Button compact floated="right">Trade</Button>
-    <Button compact color="green" floated="right">List</Button>
-    
-  </a>
-  
-)
 
 function App() {
   return (
@@ -22,55 +13,7 @@ function App() {
       <h1>Your Wallet</h1>
       <div class="ui divider"></div>
       <Container>
-        <div className="ui grid stackable" style ={{padding: "25px" }}>
-          <div class="four wide column"> 
-            <Card
-              image='https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/132.png'
-              header='Ditto'
-              meta='Normal'
-              description='Elliot is a sound engineer living in Nashville who enjoys playing guitar and hanging with his cat.'
-              extra={extra}
-            />
-          </div>
-          <div class="four wide column">
-            <Card
-                image='https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/132.png'
-                header='Ditto'
-                meta='Normal'
-                description='Elliot is a sound engineer living in Nashville who enjoys playing guitar and hanging with his cat.'
-                extra={extra}
-              />
-          </div>
-          <div class="four wide column">
-            <Card
-                image='https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/132.png'
-                header='Ditto'
-                meta='Normal'
-                description='Elliot is a sound engineer living in Nashville who enjoys playing guitar and hanging with his cat.'
-                extra={extra}
-              />
-          </div>
-          <div class="four wide column">
-            <Card
-                image='https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/132.png'
-                header='Ditto'
-                meta='Normal'
-                description='Elliot is a sound engineer living in Nashville who enjoys playing guitar and hanging with his cat.'
-                extra={extra}
-              />
-          </div>
-          <div class="four wide column">
-            <Card
-                image='https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/132.png'
-                header='Ditto'
-                meta='Normal'
-                description='Elliot is a sound engineer living in Nashville who enjoys playing guitar and hanging with his cat.'
-                extra={extra}
-              />
-          </div>
-          
-        </div>
-        
+        <Wallet/>
       </Container>
     </div>
   );
